Add monitor calibration.
Calibration checks the byte order and the duration of each call
so that we can find the exact moment an event was issued. This
suppose that optimisations are turned on (-Os).

There might be an error of one or two cycles because of the different
ways arguments are put on registers. And also because of extra
instruction to fetch the function pointer.